Run tests without subprocess and split tests and examples
