+++ .travis.yml
@@ -19,6 +19,7 @@ script:
   - npm run lint
   - npm run pretest
   - npm run test:coverage
+  - npm run test:integration
   - npm run posttest
 after_success:
   - test $TRAVIS_NODE_VERSION = "11" && npm run report-coverage
