+++ package.json
@@ -21,6 +21,7 @@
     "posttest": "node scripts/teardown-integration-tests.js",
     "test": "lerna run --parallel test",
     "test:coverage": "lerna run --parallel test:coverage",
+    "test:integration": "lerna run --parallel test:integration",
     "report-coverage": "codecov",
     "release": "lerna version",
     "publish:nightly": "lerna publish from-package --dist-tag nightly --yes"
