+++ packages/toolkit/package.json
@@ -17,7 +17,8 @@
     "test:clean": "rm -rf ./.tmp && npm run test:reset-cache",
     "test:reset-cache": "ava --reset-cache",
     "test:update-snapshots": "npm run test:clean && npm run test -- --update-snapshots",
-    "test:coverage": "nyc --all --reporter text --reporter text-summary --reporter lcovonly  npm run test"
+    "test:coverage": "nyc --all --reporter text --reporter text-summary --reporter lcovonly  npm run test",
+    "test:integration": "ava --verbose ./test-examples/*.test.js"
   },
   "keywords": [
     "aragon",
